backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.1
psycopg2-binary>=2.9.9
shap>=0.45.0
python-dotenv>=1.0.1
slowapi==0.1.9
email-validator>=2.1.0
sqlalchemy>=2.0.28
joblib>=1.3.2
